Feat/상품서비스 상품관리 웹레이어 구현

+++ src/main/java/com/yeonieum/productservice/domain/product/service/customerservice/ProductMangementServiceImpl.java
@@ -34,6 +34,17 @@ public class ProductMangementServiceImpl implements ProductManagementService{
     private final ProductCategoryRepository productCategoryRepository;
     private final SaleTypeRepository saleTypeRepository;
     private final ProductDetailImageRepository productDetailImageRepository;
+
+    /**
+     * 친환경 인증서 시리얼넘버 유효성 검증 메서드
+     * @param serialNumber
+     * @return
+     */
+    @Override
+    public boolean checkCertificationValidation(Long serialNumber) {
+        return serialNumber % 2 ==0;
+    }
+
     /**
      * 고객의 상품 등록(일반상품, 친환경상품 공통 메서드)
      *
@@ -44,7 +55,7 @@ public class ProductMangementServiceImpl implements ProductManagementService{
      */
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public boolean registerProduct(ProductManagementRequest.RegisterDto registerRequestDto, String imageUrl) {
+    public boolean registerProduct(ProductManagementRequest.RegisterDto registerRequestDto, String imageUrl) throws RuntimeException {
         try {
             Customer customer =
                     customerRepository.findById(registerRequestDto.getCustomerId()).orElseThrow(
@@ -108,10 +119,15 @@ public class ProductMangementServiceImpl implements ProductManagementService{
      * @return
      */
     @Override
-    public boolean deleteProduct(Long productId) {
+    public boolean deleteProduct(Long productId, Long customerId) {
         Product product = productRepository.findById(productId).orElseThrow(
                 () -> new IllegalArgumentException("해당하는 상품이 존재하지 않습니다.")
         );
+
+        // delete 매핑이라 customerId가 product의 소유자인지 검증
+        if(product.getCustomer().getCustomerId() != customerId) {
+            throw new IllegalArgumentException("해당 상품의 소유자가 아닙니다.");
+        }
 
         productRepository.deleteById(productId);
         return false;
@@ -127,10 +143,14 @@ public class ProductMangementServiceImpl implements ProductManagementService{
      */
     @Override
     @Transactional
-    public boolean modifyProduct(Long productId, ProductManagementRequest.ModifyDto modifyDto) {
+    public boolean modifyProduct(Long productId,  Long customerId, ProductManagementRequest.ModifyDto modifyDto) {
         Product product = productRepository.findById(productId).orElseThrow(
                 () -> new IllegalArgumentException("해당하는 상품이 존재하지 않습니다.")
         );
+        // put매핑이라 customerId가 product의 소유자인지 검증
+        if(product.getCustomer().getCustomerId() != customerId) {
+            throw new IllegalArgumentException("해당 상품의 소유자가 아닙니다.");
+        }
 
         product.changeProductName(modifyDto.getProductName());
         product.changeProductDescription(modifyDto.getDescription());
@@ -150,9 +170,9 @@ public class ProductMangementServiceImpl implements ProductManagementService{
      * @return
      */
     @Override
-    public List<ProductManagementResponse.RetrieveDto> retrieveCustomersProducts(Long customerId) {
+    public List<ProductManagementResponse.RetrieveDto> retrieveCustomersProducts(Long customerId, ActiveStatus isEcoFriend) {
         Customer customer =
-                customerRepository.findById(customerId).orElseThrow(
+                customerRepository.findCustomersWithProductsByStatus(customerId, isEcoFriend.getCode()).orElseThrow(
                         () -> new IllegalArgumentException("존재하지않는 고객 요청입니다.")
                 );
 
@@ -222,56 +242,78 @@ public class ProductMangementServiceImpl implements ProductManagementService{
      * @return
      */
     @Override
+    @Transactional(rollbackFor = Exception.class)
     public boolean uploadProductImageUrl(Long productId, String imageUrl) {
-        Product product = productRepository.findById(productId).orElseThrow(
-                () -> new IllegalArgumentException("해당하는 상품이 존재하지 않습니다.")
-        );
+        try {
+            Product product = productRepository.findById(productId).orElseThrow(
+                    () -> new IllegalArgumentException("해당하는 상품이 존재하지 않습니다.")
+            );
 
-        ProductDetailImage productDetailImage = productDetailImageRepository.findByProductId(productId);
-        if(productDetailImage != null) {
-            // 등록된 기본이미지가 있을 경우 덮어쓰기
-            productDetailImage.changeImageDetailName(imageUrl);
-        } else {
-            // 등록된 기본이미지가 없을 경우 생성
-            productDetailImage = ProductDetailImage.builder()
-                    .product(product)
-                    .imageDetailName(imageUrl)
-                    .build();
+            String defaultImageUrl = product.getProductImage();
+            product.changeProductImage(imageUrl);
+            productRepository.save(product);
+            return true;
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new RuntimeException();
         }
-
-        productDetailImageRepository.save(productDetailImage);
-        return true;
     }
 
     /**
      * 상품상세이미지등록
      * @param productId
-     * @param imageUrl
+     * @param
      * @return
      */
     @Override
-    public boolean uploadProductDetailImages(Long productId, String imageUrl) {
+    @Transactional
+    public boolean uploadProductDetailImages(Long productId,
+                                             List<Long> deleteList,
+                                             List<String> imageUrlList) {
         Product product = productRepository.findById(productId).orElseThrow(
                 () -> new IllegalArgumentException("해당하는 상품이 존재하지 않습니다.")
         );
 
-        ProductDetailImage productDetailImage = ProductDetailImage.builder()
-                .imageDetailName(imageUrl)
-                .product(product)
-                .build();
+        for(Long detailImageIdList : deleteList) {
+            ProductDetailImage productDetailImage = productDetailImageRepository.findById(detailImageIdList).orElseThrow(
+                    () -> new IllegalArgumentException("해당하는 상세이미지가 존재하지 않습니다.")
+            );
 
-        productDetailImageRepository.save(productDetailImage);
+            productDetailImageRepository.deleteById(detailImageIdList);
+        }
+
+        for(String imageUrl : imageUrlList) {
+            ProductDetailImage productDetailImage = ProductDetailImage.builder()
+                    .imageDetailName(imageUrl)
+                    .product(product)
+                    .build();
+
+            productDetailImageRepository.save(productDetailImage);
+        }
+
         return true;
     }
 
-
-    public boolean deleteProductDetailImage(Long productDetailImageId) {
-        ProductDetailImage productDetailImage = productDetailImageRepository.findById(productDetailImageId).orElseThrow(
-                () -> new IllegalArgumentException("해당하는 상세이미지가 존재하지 않습니다.")
+    @Override
+    public List<ProductManagementResponse.RetrieveDetailImage> retrieveDetailImage(Long productId) {
+        Product product = productRepository.findById(productId).orElseThrow(
+                () -> new IllegalArgumentException("해당하는 상품이 존재하지 않습니다.")
         );
 
-        productDetailImageRepository.deleteById(productDetailImageId);
-        return true;
+        List<ProductDetailImage> productDetailImageList = productDetailImageRepository.findByProductId(productId);
+
+        List<ProductManagementResponse.RetrieveDetailImage> detailImageList =
+                productDetailImageList.stream().map(
+                        productDetailImage -> {
+                            return ProductManagementResponse.RetrieveDetailImage.builder()
+                                    .imageName(productDetailImage.getImageDetailName())
+                                    .productId(productId)
+                                    .productDetailImageId(productDetailImage.getProductDetailImageId())
+                                    .build();
+                        }
+                ).collect(Collectors.toList());
+
+        return detailImageList;
     }
 
 
@@ -306,6 +348,7 @@ public class ProductMangementServiceImpl implements ProductManagementService{
         return true;
     }
 
+    @Transactional
     public void registerCertificationOf(Product product, ProductManagementRequest.RegisterDto registerRequestDto) {
         ProductManagementRequest.Certification certification
                 = ((ProductManagementRequest.RegisterEcoFriendlyProductDto) registerRequestDto).getCertification();
